+++ unit/test_scenario.py
@@ -1,22 +1,22 @@
 import pytest
 from unittest.mock import patch, MagicMock
-from src.scenario import Scenario
 from src.file import File
 from src.response import Response
 from src.config import output_path, files_path, expected_files_path
 from src.exceptions import GitExecutionError
+from src.scenario import Scenario
 
 
 @pytest.fixture
 def scenario():
-    return Scenario(name="test.scenario", remote=True)
+    return Scenario(name="test.scenario")
 
 
 @patch("src.scenario.copyfile")
 def test_add_file(mock_copyfile, scenario):
-    scenario.scenario_local_path = output_path / "test_scenario" / "local"
+    scenario.local_path = output_path / "test_scenario" / "local"
     scenario.add_file("source.txt", "dest.txt")
-    mock_copyfile.assert_called_once_with(src=files_path / "source.txt", dst=scenario.scenario_local_path / "dest.txt")
+    mock_copyfile.assert_called_once_with(src=files_path / "source.txt", dst=scenario.local_path / "dest.txt")
     assert "dest.txt" in scenario._files
     assert isinstance(scenario._files["dest.txt"], File)
 
@@ -30,7 +30,7 @@ def test_get_file(scenario):
 
 def test_get_expected_file():
     file_name = "expected.txt"
-    expected_file = Scenario.get_expected_file(file_name)
+    expected_file = ScenarioOld.get_expected_file(file_name)
     assert isinstance(expected_file, File)
     assert expected_file.path == expected_files_path / file_name
 
@@ -41,7 +41,7 @@ def test_run_success(mock_run, scenario):
     command = "git status"
     response = scenario.run(command)
     mock_run.assert_called_once_with(
-        ["git", "status"], cwd=scenario.scenario_local_path, capture_output=True, text=True
+        ["git", "status"], cwd=scenario.local_path, capture_output=True, text=True
     )
     assert isinstance(response, Response)
 
@@ -60,6 +60,6 @@ def test_run_failure_exception(mock_run, scenario):
     command = "git status"
     response = scenario.run(command)
     mock_run.assert_called_once_with(
-        ["git", "status"], cwd=scenario.scenario_local_path, capture_output=True, text=True
+        ["git", "status"], cwd=scenario.local_path, capture_output=True, text=True
     )
     assert isinstance(response, Response)